Correct Finnish translation for "min"
"Min" has a double meaning in Finnish. It used to be translated as "Matalin" which actually means "lowest", often referred as "min" (in min-max). Yet in the UI I believe this is only used in Media information section to indicate the duration of a video. So it's actually used as a shortcut for "minutes" and as such should be translated as "min".